+++ test/dummy.test.js
@@ -22,7 +22,7 @@ describe("dummy", () => {
     expect(response.body).toBeDefined();
     expect(response.body).toEqual({
       d: {
-        EntitySets: ["Header", "HeaderItem", "HeaderLine"],
+        EntitySets: ["Header", "HeaderItem", "HeaderLine", "Header_texts", "HeaderItem_texts"],
       },
     });
   });