+++ test/protocols.test.js
@@ -23,7 +23,7 @@ const expectGETService = async (request, path) => {
   expect(response.body).toBeDefined();
   expect(response.body).toEqual({
     d: {
-      EntitySets: ["Header", "HeaderItem", "HeaderLine"],
+      EntitySets: ["Header", "HeaderItem", "HeaderLine", "Header_texts", "HeaderItem_texts"],
     },
   });
 };